Move model mapping into model extensions using helper methods
