+++ Tests/ManagedCache.swift
@@ -9,9 +9,22 @@
 
 import Foundation
 import CoreData
+import FeedStoreChallenge
 
 @objc(ManagedCache)
 public class ManagedCache: NSManagedObject {
     @NSManaged var timestamp: Date
     @NSManaged var feed: NSOrderedSet
+}
+
+extension ManagedCache {
+    var localFeed: [LocalFeedImage] {
+        return feed
+            .compactMap {
+                guard let managedFeedImage = $0 as? ManagedFeedImage else {
+                    return nil
+                }
+                return managedFeedImage.localFeedImage
+        }
+    }
 }
