+++ Tests/ManagedFeedImage.swift
@@ -9,6 +9,7 @@
 
 import Foundation
 import CoreData
+import FeedStoreChallenge
 
 @objc(ManagedFeedImage)
 public class ManagedFeedImage: NSManagedObject {
@@ -17,4 +18,22 @@ public class ManagedFeedImage: NSManagedObject {
     @NSManaged public var location: String?
     @NSManaged public var url: URL
     @NSManaged public var cache: ManagedCache
+}
+
+extension ManagedFeedImage {
+    static func feed(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
+        return NSOrderedSet(array: localFeed.map {
+            let managedFeedImage = ManagedFeedImage(context: context)
+            managedFeedImage.id = $0.id
+            managedFeedImage.imageDescription = $0.description
+            managedFeedImage.location = $0.location
+            managedFeedImage.url = $0.url
+            
+            return managedFeedImage
+        })
+    }
+    
+    var localFeedImage: LocalFeedImage {
+        return LocalFeedImage(id: id, description: imageDescription, location: location, url: url)
+    }
 }
